+++ intervention_graph_creation/src/local_graph_extraction/core/node.py
@@ -1,5 +1,6 @@
-from typing import List, Optional, Literal, Union
+from typing import Optional, Literal, List
 from pydantic import BaseModel, Field, ConfigDict, field_validator, model_validator
+import numpy as np
 
 
 class Node(BaseModel):
@@ -58,65 +59,13 @@ class Node(BaseModel):
         return self
 
 
-class Edge(BaseModel):
-    type: str = Field(min_length=1, max_length=64, description="relationship label verb")
-    source_node: str = Field(min_length=1, description="source node name")
-    target_node: str = Field(min_length=1, description="target node name")
-    description: str = Field(min_length=1, description="concise description of logical connection")
-    edge_confidence: int = Field(ge=1, le=5, description="1-5")
+class GraphNode(Node):
+    """Extended Node class with embedding support."""
+    embedding: Optional[np.ndarray] = None
+    model_config = ConfigDict(arbitrary_types_allowed=True)
 
-    model_config = ConfigDict(extra="forbid")
-
-    @field_validator("type", "source_node", "target_node", "description")
-    @classmethod
-    def _strip_nonempty(cls, v: str) -> str:
-        v2 = v.strip()
-        if not v2:
-            raise ValueError("must be non-empty")
-        return v2
-
-    @model_validator(mode="after")
-    def _no_self_loop(self):
-        if self.source_node == self.target_node:
-            raise ValueError("self-loop edges are not allowed (source_node == target_node)")
-        return self
-    
-
-class Meta(BaseModel):
-    key: str = Field(min_length=1, max_length=64, description="metadata key")
-    value: Union[str, List[str]] = Field(min_length=1, max_length=256, description="metadata value")
-
-    model_config = ConfigDict(extra="forbid")
-
-    @field_validator("key", "value", mode="before")
-    @classmethod
-    def _strip_nonempty(cls, v):
-        if isinstance(v, str):
-            v2 = v.strip()
-            if not v2:
-                raise ValueError("must be non-empty")
-            return v2
-        elif isinstance(v, list):
-            new_list = []
-            for item in v:
-                if not isinstance(item, str):
-                    raise ValueError("all items must be strings")
-                item2 = item.strip()
-                if not item2:
-                    raise ValueError("list items must be non-empty")
-                new_list.append(item2)
-            return new_list
-        return v
-
-
-class LogicalChain(BaseModel):
-    title: Optional[str] = Field(default=None, description="concise natural-language description of logical chain")
-    edges: List[Edge] = Field(default_factory=list)
-    model_config = ConfigDict(extra="forbid")
-
-
-class PaperSchema(BaseModel):
-    nodes: List[Node] = Field(default_factory=list)
-    logical_chains: List[LogicalChain] = Field(default_factory=list)
-    meta: List[Meta] = Field(default_factory=list)
-    model_config = ConfigDict(extra="forbid")
+    def __init__(self, **data):
+        # Handle embedding separately to avoid pydantic validation issues
+        embedding = data.pop('embedding', None)
+        super().__init__(**data)
+        self.embedding = embedding
